aider: Added pencil and trash icons for rename and delete functionality in Sidebar.
# Aider chat conversation:

USER: the pencil icon and trash icon for rename and delete functionality are missing

ASSISTANT: To add the pencil and trash icons for rename and delete functionality, we need to update the `src/components/Sidebar.js` file. We'll use Font Awesome icons for this purpose.

Here are the necessary changes:

src/components/Sidebar.js
```javascript
<<<<<<< SEARCH
              </>
=======
                <i
                  className="fas fa-pencil-alt ml-2"
                  onClick={() => {
                    setRenamingChat(index);
                    setRenameValue(chat.name);
                  }}
                ></i>
                <i
                  className="fas fa-trash ml-2"
                  onClick={() => {
                    const updatedChats = chats.filter((_, i) => i !== index);
                    setChats(updatedChats);
                  }}
                ></i>
              </>
>>>>>>> REPLACE
```

This will add the pencil and trash icons next to each chat name, allowing users to rename or delete chats.